color_extractor_web_app/requirements.txt
Flask==2.3.2
numpy==1.24.4
scikit-learn==1.3.0
Pillow==10.0.0
matplotlib==3.7.1